+++ src/components/Footer.jsx
@@ -4,10 +4,11 @@ import { FaGithub } from 'react-icons/fa';
 const Footer = () => {
   return (
     <div className='footer'>
-      <p className="footer__text">
-        <i>XIV Tracker</i> is Designed and Developed by Damon Greenhalgh.<br/>
-        FINAL FANTASY XIV CONTENT IS PROPERTY OF SQUARE ENIX CO,. LTD.
+      <p>
+        <b>XIV Tracker</b> is Designed and Developed by Damon Greenhalgh.<br/>
+        <b>FINAL FANTASY XIV CONTENT IS PROPERTY OF SQUARE ENIX CO,. LTD.</b>
       </p>
+      <p style={{marginLeft: 'auto', minWidth: 'max-content'}}><b>XIV Tracker v1.2.1</b></p>
       <button>
         <a href="https://github.com/DamonGreenhalgh/xivtracker">
           <FaGithub size="1.25em" />