+++ src/components/Navbar.jsx
@@ -23,24 +23,16 @@ const Navbar = (props) => {
             </Link>
             {
                 props.showSearchbar ?
-                <Searchbar search={(name, server) => window.location.href = "../?name=" + name + "&server=" + server} /> : 
+                <Searchbar 
+                    search={(name, server) => window.location.href = "./?name=" + name + "&server=" + server} 
+                    type='character'
+                /> : 
                 null
             }
             <Link to="/settings" title="Settings" style={{marginLeft: "auto"}}>
                 <button><RiSettings3Line className="navbar__icon" /></button>
             </Link>
             {profile}
-            {/* {
-                 props.referenceCharacter !== null ?
-                <Link to={"/" + props.referenceCharacter.Character.ID}>
-                    <img 
-                        src={props.referenceCharacter.Character.Avatar} 
-                        className="navbar__profile-icon rounded interactable" 
-                        title={props.referenceCharacter.Character.Name}
-                    />
-                </Link> :
-                null
-            } */}
         </nav>
     );
 }